fix(editor): apply unload confirm to setCover
