+++ lang/zh-Hans.json
@@ -1126,6 +1126,9 @@
     "defaultMessage": "你已拥有个人创作空间站，期待你的第一篇作品。",
     "description": "src/components/Forms/EmailSignUpForm/Complete.tsx"
   },
+  "rfz/fN": {
+    "defaultMessage": "建議尺寸 1600 x 900 像素"
+  },
   "s5JCCO": {
     "defaultMessage": "此文已被{actor}标示为限制级内容，是否确定展开全文？",
     "description": "src/views/ArticleDetail/Wall/Sensitive/index.tsx"
