+++ lang/zh-Hant.json
@@ -1126,6 +1126,9 @@
     "defaultMessage": "你已擁有個人創作空間站，期待你的第一篇作品。",
     "description": "src/components/Forms/EmailSignUpForm/Complete.tsx"
   },
+  "rfz/fN": {
+    "defaultMessage": "建議尺寸 1600 x 900 像素"
+  },
   "s5JCCO": {
     "defaultMessage": "此文已被{actor}標示為限制級內容，是否確定展開全文？",
     "description": "src/views/ArticleDetail/Wall/Sensitive/index.tsx"
