+++ src/components/Hook/index.ts
@@ -1,3 +1,4 @@
+export * from './useBlockUnload'
 export * from './useCache'
 export * from './useCarousel'
 export * from './useColorThief'
