+++ src/views/Me/DraftDetail/index.tsx
@@ -1,7 +1,7 @@
 import { useQuery } from '@apollo/react-hooks'
 import dynamic from 'next/dynamic'
-import { useEffect, useState } from 'react'
-import { FormattedMessage, useIntl } from 'react-intl'
+import { useState } from 'react'
+import { FormattedMessage } from 'react-intl'
 
 import {
   ASSET_TYPE,
@@ -17,6 +17,7 @@ import {
   Spinner,
   Throw404,
   toast,
+  useBlockUnload,
   useRoute,
 } from '~/components'
 import { QueryError, useMutation } from '~/components/GQL'
@@ -43,8 +44,7 @@ const Editor = dynamic(
 )
 
 const DraftDetail = () => {
-  const intl = useIntl()
-  const { getQuery, router } = useRoute()
+  const { getQuery } = useRoute()
   const id = getQuery('draftId')
 
   const [setContent] = useMutation<SetDraftContentMutation>(SET_CONTENT)
@@ -64,45 +64,7 @@ const DraftDetail = () => {
   const draft = (data?.node?.__typename === 'Draft' && data.node) || undefined
   const ownCircles = data?.viewer?.ownCircles || undefined
 
-  const onLeaveConfirmText = intl.formatMessage({
-    defaultMessage: 'Saving draft, are you sure you want to leave?',
-  })
-
-  const onRouteChange = () => {
-    if (saveStatus !== 'saving') {
-      return null
-    }
-
-    if (window.confirm(onLeaveConfirmText)) {
-      return true
-    }
-
-    throw "Abort route change by user's confirmation."
-  }
-
-  const onBeforeUnload = (event: any) => {
-    if (saveStatus !== 'saving') {
-      return null
-    }
-
-    // legacy browsers
-    if (event) {
-      event.returnValue = onLeaveConfirmText
-    }
-
-    // modern browsers (ignore text)
-    return onLeaveConfirmText
-  }
-
-  useEffect(() => {
-    window.addEventListener('beforeunload', onBeforeUnload)
-    router.events.on('routeChangeStart', onRouteChange)
-
-    return () => {
-      window.removeEventListener('beforeunload', onBeforeUnload)
-      router.events.off('routeChangeStart', onRouteChange)
-    }
-  }, [])
+  useBlockUnload({ isBlocking: saveStatus === 'saving' })
 
   if (loading) {
     return (
